airsense-backend/ml/requirements.txt
pymongo>=4.6
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
python-dotenv>=1.0
joblib>=1.3
